+++ tab.go
@@ -1,5 +1,9 @@
 package readline
 
+import (
+	"context"
+)
+
 // TabDisplayType defines how the autocomplete suggestions display
 type TabDisplayType int
 
@@ -25,10 +29,18 @@ func (rl *Instance) getTabCompletion() {
 		return
 	}
 
-	rl.tcPrefix, rl.tcSuggestions, rl.tcDescriptions, rl.tcDisplayType = rl.TabCompleter(rl.line, rl.pos)
-	if len(rl.tcSuggestions) == 0 {
-		return
+	if rl.delayedTabContext.cancel != nil {
+		rl.delayedTabContext.cancel()
 	}
+
+	rl.delayedTabContext = DelayedTabContext{rl: rl}
+	rl.delayedTabContext.Context, rl.delayedTabContext.cancel = context.WithCancel(context.Background())
+
+	rl.tcPrefix, rl.tcSuggestions, rl.tcDescriptions, rl.tcDisplayType = rl.TabCompleter(rl.line, rl.pos, rl.delayedTabContext)
+	/*if len(rl.tcSuggestions) == 0 && delayed {
+		return
+	}*/
+	//panic(rl.tcDisplayType)
 
 	if len(rl.tcDescriptions) == 0 {
 		// probably not needed, but just in case someone doesn't initialise the
@@ -63,10 +75,15 @@ func (rl *Instance) moveTabCompletionHighlight(x, y int) {
 	}
 }
 
-func (rl *Instance) writeTabCompletion() {
+func (rl *Instance) writeTabCompletion(resetCursorPos bool) {
 	if !rl.modeTabCompletion {
 		return
 	}
+
+	_, posY := lineWrapPos(rl.promptLen, rl.pos, rl.termWidth)
+	_, lineY := lineWrapPos(rl.promptLen, len(rl.line), rl.termWidth)
+	moveCursorDown(rl.hintY + lineY - posY)
+	print("\r\n" + seqClearScreenBelow)
 
 	switch rl.tcDisplayType {
 	case TabDisplayGrid:
@@ -80,6 +97,12 @@ func (rl *Instance) writeTabCompletion() {
 
 	default:
 		rl.writeTabGrid()
+	}
+
+	if resetCursorPos {
+		moveCursorUp(rl.hintY + rl.tcUsedY)
+		print("\r")
+		rl.moveCursorFromStartToLinePos()
 	}
 }